// src/types.ts
export type Plant = {
  id: string;
  scientific_name: string;
  rank?: string;
  //   genus: string;
  //   family: string;
  //   year: string;
  //   author: string;
  //   bibliography: string;
  common_name?: string;
  family_common_name?: string;
    image_url: string;
  //   flower_color: string;
  //   flower_conspicuous: string;
  //   foliage_color: string;
  //   foliage_texture: string;
  //   fruit_color: string;
  //   fruit_conspicuous: string;
  //   fruit_months: string;
  //   bloom_months: string;
  //   ground_humidity: string;
  //   growth_form: string;
  //   growth_habit: string;
  //   growth_months: string;
  //   growth_rate: string;
  //   edible_part: string;
  //   vegetable: boolean;
  //   edible: boolean;
  //   light: string;
  //   soil_nutriments: string;
  //   soil_salinity: string;
  //   anaerobic_tolerance: string;
  //   atmospheric_humidity: string;
  //   average_height_cm: string;
  //   maximum_height_cm: string;
  //   minimum_root_depth_cm: string;
  //   ph_maximum: string;
  //   ph_minimum: string;
  //   planting_days_to_harvest: string;
  //   planting_description: string;
  //   planting_sowing_description: string;
  //   planting_row_spacing_cm: string;
  //   planting_spread_cm: string;
  //   synonyms: string;
  //   distributions: string;
  common_names?: string;
  //   url_usda:string;
  //   url_tropicos: string;
  //   url_tela_botanica: string;
  //   url_powo: string;
  //   url_plantnet: string;
  //   url_gbif: string;
  //   url_openfarm: string;
  //   url_catminat: string;
  url_wikipedia_en: string;
};